requirements.txt
aiogram
python-dotenvпше